requirements.txt
altair==5.5.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.15.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-assistant==0.4.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
protobuf==5.29.1
pyarrow==18.1.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.1
streamlit==1.40.2
streamlit-lottie==0.0.5
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
